Enable Vcpkg build (#26)
* Code fixes for MSVC errors

* Additional changes to build with vcpkg

* Put in stupid windows solution for now.

* Woops.

* Only load file-by-file once.

* Update CMakeLists.txt

* Create vcpkg.json

* Create windows_cmake.yml

* Update windows_cmake.yml

* Update windows_cmake.yml

* Update windows_cmake.yml

* Create CMakePresets.json

* Update windows_cmake.yml

* Update windows_cmake.yml

* Update windows_cmake.yml

* Update vcpkg.json

* Remove CMakePresets.json, and remove BOOST_DEBUG

* Add configure args to use vcpkg--need to set this only for windows

* Remove clang tidy when builing with MSVC

* Reduce code duplication in CMakeLists.txt

* Remove macos from github actions

* Conditionally add pthread

* Add CMakePresets.json

* Corrected names

* Add generator description to WSL presets

* Update windows_cmake.yml

+++ src/new_client/interfacelib/PlayerEquipmentDisplay.cpp
@@ -64,12 +64,12 @@ void PlayerEquipmentDisplay::onUserInput( const sf::Event& e )
 
     if ( MenAmongGods::equipmentBoundingBox.contains( mousePosition ) )
     {
-      int itemRow = std::floor( ( mousePosition.y - MenAmongGods::equipmentBoundingBoxPosition.y ) /
-                                ( MenAmongGods::equipmentBoundingBox.height / 6.0f ) );
+      int itemRow = static_cast< int >( std::floor( ( mousePosition.y - MenAmongGods::equipmentBoundingBoxPosition.y ) /
+                                                    ( MenAmongGods::equipmentBoundingBox.height / 6.0f ) ) );
 
       // We'll do a similar strategy for the columns, except there are only two columns in this case
-      int itemColumn = std::floor( ( mousePosition.x - MenAmongGods::equipmentBoundingBoxPosition.x ) /
-                                   ( MenAmongGods::equipmentBoundingBox.width / 2.0f ) );
+      int itemColumn = static_cast< int >( std::floor( ( mousePosition.x - MenAmongGods::equipmentBoundingBoxPosition.x ) /
+                                                       ( MenAmongGods::equipmentBoundingBox.width / 2.0f ) ) );
 
       // TODO: The last parameter should be the selected character--though
       // not sure why that would matter for equipping things...
@@ -130,12 +130,12 @@ void PlayerEquipmentDisplay::onUserInput( const sf::Event& e )
 
     if ( MenAmongGods::equipmentBoundingBox.contains( mousePosition ) )
     {
-      int itemRow = std::floor( ( mousePosition.y - MenAmongGods::equipmentBoundingBoxPosition.y ) /
-                                ( MenAmongGods::equipmentBoundingBox.height / 6.0f ) );
+      int itemRow = static_cast< int >( std::floor( ( mousePosition.y - MenAmongGods::equipmentBoundingBoxPosition.y ) /
+                                                    ( MenAmongGods::equipmentBoundingBox.height / 6.0f ) ) );
 
       // We'll do a similar strategy for the columns, except there are only two columns in this case
-      int itemColumn = std::floor( ( mousePosition.x - MenAmongGods::equipmentBoundingBoxPosition.x ) /
-                                   ( MenAmongGods::equipmentBoundingBox.width / 2.0f ) );
+      int itemColumn = static_cast< int >( std::floor( ( mousePosition.x - MenAmongGods::equipmentBoundingBoxPosition.x ) /
+                                                       ( MenAmongGods::equipmentBoundingBox.width / 2.0f ) ) );
 
       // TODO: The last parameter should be the selected character--though
       // not sure why that would matter for equipping things...
